Odradjeno odobravanje komentara na agentskom modulu

+++ backend/src/main/java/com/project/Rentingaccommodation/controller/CommentController.java
@@ -14,11 +14,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.project.Rentingaccommodation.model.Admin;
 import com.project.Rentingaccommodation.model.Apartment;
 import com.project.Rentingaccommodation.model.Comment;
+import com.project.Rentingaccommodation.model.CommentStatus;
 import com.project.Rentingaccommodation.model.User;
 import com.project.Rentingaccommodation.security.JwtUser;
 import com.project.Rentingaccommodation.security.JwtValidator;
+import com.project.Rentingaccommodation.service.AdminService;
 import com.project.Rentingaccommodation.service.ApartmentService;
 import com.project.Rentingaccommodation.service.CommentService;
 import com.project.Rentingaccommodation.service.UserService;
@@ -37,6 +40,9 @@ public class CommentController {
     private ApartmentService apartmentService;
     
     @Autowired
+    private AdminService adminService;
+    
+    @Autowired
     private JwtValidator jwtValidator;
     
 	@RequestMapping(value="", method=RequestMethod.GET)
@@ -47,11 +53,6 @@ public class CommentController {
 	@RequestMapping(value="/approved", method=RequestMethod.GET)
     public ResponseEntity<List<Comment>> getApprovedComments() {
     	return new ResponseEntity<>(service.findApprovedComments(), HttpStatus.OK);
-    }
-	
-	@RequestMapping(value="/declined", method=RequestMethod.GET)
-    public ResponseEntity<List<Comment>> getDeclinedComments() {
-    	return new ResponseEntity<>(service.findDeclinedComments(), HttpStatus.OK);
     }
 	
 	@RequestMapping(value="/waiting", method=RequestMethod.GET)
@@ -71,6 +72,11 @@ public class CommentController {
 	@RequestMapping(value="/apartment/{id}", method=RequestMethod.GET)
     public List<Comment> getAllCommentsForApartment(@PathVariable Long id) {
         return service.findApartmentComments(id);
+    }
+	
+	@RequestMapping(value="/apartment/{id}/approved", method=RequestMethod.GET)
+    public List<Comment> getAllApprovedCommentsForApartment(@PathVariable Long id) {
+        return service.findApartmentApprovedComments(id);
     }
     
 	@RequestMapping(value="", method=RequestMethod.POST)
@@ -99,6 +105,16 @@ public class CommentController {
 				String date = dateFormatter.format(new Date());
 				String time = timeFormatter.format(new Date());
 				
+				for (Comment c : service.findAll()) {
+					if (c.getApartment().getId() == apartment.getId() && c.getUser().getId() == user.getId()) {
+						if (c.getStatus().equals(CommentStatus.APPROVED)) {
+							return new ResponseEntity<>("User already commented this apartment.", HttpStatus.FORBIDDEN);
+						} else {
+							return new ResponseEntity<>("User already commented this apartment. Waiting for approval.", HttpStatus.FORBIDDEN);
+						}
+					}
+				}
+				
 				Comment comment = new Comment(user, apartment, data.getComment(), date, time);
 				
 		        return new ResponseEntity<>(service.save(comment), HttpStatus.OK);
@@ -106,22 +122,55 @@ public class CommentController {
 				return new ResponseEntity<>("User with this email doesn't exist.", HttpStatus.NOT_FOUND);
 			}
 		} catch (Exception e) {
-			return new ResponseEntity<>("Token not provided.", HttpStatus.FORBIDDEN);
+			return new ResponseEntity<>("Error validating token.", HttpStatus.FORBIDDEN);
 		}
     }
 
-	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-    public ResponseEntity<Object> deleteComment(@PathVariable Long id) {
-		Comment comment = service.findOne(id);
-		if (comment == null) {
-			return new ResponseEntity<>("Comment not found.", HttpStatus.NOT_FOUND);
+	@RequestMapping(value="/{id}/approve", method=RequestMethod.PUT)
+    public ResponseEntity<Object> approveComment(@PathVariable Long id, @RequestHeader("Authorization") String authHeader) {
+		try {
+			String token = authHeader.split(" ")[1].trim();
+			JwtUser jwtUser = jwtValidator.validate(token);
+			if (jwtUser != null) {
+				Admin admin = adminService.findOne(jwtUser.getId());
+				Comment comment = service.findOne(id);
+				if (admin == null) {
+					return new ResponseEntity<>("Admin not found.", HttpStatus.NOT_FOUND);
+				}
+				if (comment == null) {
+					return new ResponseEntity<>("Comment not found.", HttpStatus.NOT_FOUND);
+				}
+				comment.setStatus(CommentStatus.APPROVED);
+				return new ResponseEntity<>(service.save(comment), HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>("User with this email doesn't exist.", HttpStatus.NOT_FOUND);
+			}
+		} catch (Exception e) {
+			return new ResponseEntity<>("Error validating token.", HttpStatus.FORBIDDEN);
 		}
-		service.delete(comment);
-        return new ResponseEntity<>(comment, HttpStatus.OK);
     }
 	
-//	@RequestMapping(value="/accommodation/{id}/grade", method=RequestMethod.GET)
-//    public double calculateReview(@PathVariable Long id){
-//        return service.calculateAverageGrade(id);
-//    }
+	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
+    public ResponseEntity<Object> deleteComment(@PathVariable Long id, @RequestHeader("Authorization") String authHeader) {
+		try {
+			String token = authHeader.split(" ")[1].trim();
+			JwtUser jwtUser = jwtValidator.validate(token);
+			if (jwtUser != null) {
+				Admin admin = adminService.findOne(jwtUser.getId());
+				Comment comment = service.findOne(id);
+				if (admin == null) {
+					return new ResponseEntity<>("Admin not found.", HttpStatus.NOT_FOUND);
+				}
+				if (comment == null) {
+					return new ResponseEntity<>("Comment not found.", HttpStatus.NOT_FOUND);
+				}
+				service.delete(comment);
+				return new ResponseEntity<>(comment, HttpStatus.OK);
+			}  else {
+				return new ResponseEntity<>("User with this email doesn't exist.", HttpStatus.NOT_FOUND);
+			}
+		} catch (Exception e) {
+			return new ResponseEntity<>("Error validating token.", HttpStatus.FORBIDDEN);
+		}
+    }
 }
